% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnut.R
\name{get_haz}
\alias{get_haz}
\alias{get_stunt}
\title{Determining Child Malnutrition Status}
\usage{
get_haz(hv103, hc70)

get_stunt(haz)
}
\arguments{
\item{hv103}{Classification of de facto children or not variable from
PR dataset}

\item{hc70}{HAZ score variable from PR dataset}

\item{haz}{clenaed HAZ score}
}
\value{
a continious vector of haz z score
}
\description{
Identification of individual 6-59 months old children
malnutrition status
}
\examples{
  # HAZ mean score
  haz_score <- get_haz(PRdata$hv103, PRdata$hc70)

  wt <- PRdata$hv005/1000000

  weighted.mean(haz_score, wt, na.rm = TRUE)

  summary(haz_score, na.rm = TRUE)

  # stunting
  stunting <- get_stunt(haz_score)
  xtabs(wt ~ stunting)
  sum(xtabs(wt ~ stunting))
  prop.table(xtabs(wt ~ stunting))



}
\author{
Nicholus Tint Zaw
}
