% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IRdata}
\alias{IRdata}
\title{IR dataset}
\format{
A data frame with XX columns and XX rows.
}
\source{
DHS
}
\usage{
IRdata
}
\description{
Dataset description - add later
}
\examples{
# explore the variable names
names(IRdata)

}
\keyword{datasets}
