% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PRdata}
\alias{PRdata}
\title{DHS sample dataset}
\format{
A data frame with XX columns and XX rows.
}
\source{
DHS
}
\usage{
PRdata
}
\description{
PR dataset
}
\details{
Dataset description - add later
}
\examples{
# explore the de facto children
table(PRdata$hv103)

}
\keyword{datasets}
